requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
sqlalchemy>=1.4.0,<1.5.0
psycopg2-binary>=2.9.1,<3.0.0
alembic>=1.7.0,<1.8.0
python-dotenv>=0.19.0,<0.20.0
httpx>=0.24.0,<0.25.0
pydantic>=1.8.0,<2.0.0 